Switched from Integer to BigInteger types to fix issues with too small types

+++ matebot_telegram/persistence.py
@@ -6,7 +6,7 @@ import sys
 from typing import Optional
 from datetime import datetime as _dt
 
-from sqlalchemy import create_engine, Column, DateTime, FetchedValue, Integer, String
+from sqlalchemy import BigInteger, create_engine, Column, DateTime, FetchedValue, String
 from sqlalchemy.engine import Engine as _Engine
 from sqlalchemy.orm import declarative_base, sessionmaker, Session
 from sqlalchemy.sql import func
@@ -89,8 +89,8 @@ class TelegramUser(Base):
 
     __tablename__ = "telegram_users"
 
-    telegram_id: int = Column(Integer, nullable=False, primary_key=True, unique=True)
-    user_id: int = Column(Integer, nullable=False, unique=True)
+    telegram_id: int = Column(BigInteger, nullable=False, primary_key=True, unique=True)
+    user_id: int = Column(BigInteger, nullable=False, unique=True)
     first_name: str = Column(String(32), nullable=False)
     last_name: str = Column(String(64), nullable=True)
     username: str = Column(String(64), nullable=True)
@@ -108,11 +108,11 @@ class SharedMessage(Base):
 
     __tablename__ = "shared_messages"
 
-    id: int = Column(Integer, nullable=False, primary_key=True, autoincrement=True, unique=True)
+    id: int = Column(BigInteger, nullable=False, primary_key=True, autoincrement=True, unique=True)
     share_type: str = Column(String(32), nullable=False)
-    share_id: int = Column(Integer, nullable=False)
-    chat_id: int = Column(Integer, nullable=False)
-    message_id: int = Column(Integer, nullable=False)
+    share_id: int = Column(BigInteger, nullable=False)
+    chat_id: int = Column(BigInteger, nullable=False)
+    message_id: int = Column(BigInteger, nullable=False)
     created: _dt = Column(DateTime, server_default=func.now())
     modified: _dt = Column(DateTime, server_onupdate=FetchedValue(), server_default=func.now(), onupdate=func.now())
 
@@ -127,9 +127,9 @@ class RegistrationProcess(Base):
 
     __tablename__ = "registration_processes"
 
-    telegram_id: int = Column(Integer, nullable=False, primary_key=True, unique=True, autoincrement=True)
-    application_id: int = Column(Integer, nullable=False)
+    telegram_id: int = Column(BigInteger, nullable=False, primary_key=True, unique=True, autoincrement=True)
+    application_id: int = Column(BigInteger, nullable=False)
     selected_username: str = Column(String(255), nullable=True)
-    core_user_id: int = Column(Integer, nullable=True)
+    core_user_id: int = Column(BigInteger, nullable=True)
     created: _dt = Column(DateTime, server_default=func.now())
     modified: _dt = Column(DateTime, server_onupdate=FetchedValue(), server_default=func.now(), onupdate=func.now())
